Share OpenTKMouseState between both handlers
